savecode/pythonpackages/requirements.txt
requests>=2.18.4
pytz>=2019.1
pyOpenSSL>=19.0.0
pycryptodome>=3.9.8
crypto>=1.4.1